+++ X10D.Tests/src/Drawing/PointFTests.cs
@@ -9,6 +9,26 @@ namespace X10D.Tests.Drawing;
 public class PointFTests
 {
     [TestMethod]
+    public void Round_ShouldRoundToNearestInteger_GivenNoParameters()
+    {
+        var point = new PointF(1.5f, 2.6f);
+        var rounded = point.Round();
+
+        Assert.AreEqual(2, rounded.X);
+        Assert.AreEqual(3, rounded.Y);
+    }
+
+    [TestMethod]
+    public void Round_ShouldRoundToNearest10_GivenPrecision10()
+    {
+        var point = new PointF(1.5f, 25.2f);
+        var rounded = point.Round(10);
+
+        Assert.AreEqual(0, rounded.X);
+        Assert.AreEqual(30, rounded.Y);
+    }
+
+    [TestMethod]
     public void ToSizeF_ShouldReturnSize_WithEquivalentMembers()
     {
         var random = new Random();