+++ X10D.Tests/src/Numerics/Vector2Tests.cs
@@ -19,6 +19,26 @@ public class Vector2Tests
     }
 
     [TestMethod]
+    public void Round_ShouldRoundToNearestInteger_GivenNoParameters()
+    {
+        var vector = new Vector2(1.5f, 2.6f);
+        var rounded = vector.Round();
+
+        Assert.AreEqual(2, rounded.X);
+        Assert.AreEqual(3, rounded.Y);
+    }
+
+    [TestMethod]
+    public void Round_ShouldRoundToNearest10_GivenPrecision10()
+    {
+        var vector = new Vector2(1.5f, 25.2f);
+        var rounded = vector.Round(10);
+
+        Assert.AreEqual(0, rounded.X);
+        Assert.AreEqual(30, rounded.Y);
+    }
+
+    [TestMethod]
     public void ToPointF_ShouldReturnPoint_WithEquivalentMembers()
     {
         var random = new Random();