+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - X10D: Added `Point.ToSize()`
 - X10D: Added `Point.ToSizeF()`
 - X10D: Added `Point.ToVector2()`
+- X10D: Added `PointF.Round([float])`
 - X10D: Added `PointF.ToSizeF()`
 - X10D: Added `PointF.ToVector2()` for .NET < 6
 - X10D: Added `PopCount()` for built-in integer types
@@ -22,10 +23,13 @@
 - X10D: Added `Quaternion.Multiply(Vector3)` - this functions as an equivalent to Unity's `Quaternion * Vector3` operator
 - X10D: Added `Vector2.Deconstruct()`
 - X10D: Added `Vector2.IsOnLine(LineF)`, `Vector2.IsOnLine(PointF, PointF)`, and `Vector2.IsOnLine(Vector2, Vector2)`
+- X10D: Added `Vector2.Round([float])`
 - X10D: Added `Vector2.ToPointF()`
 - X10D: Added `Vector2.ToSizeF()`
 - X10D: Added `Vector3.Deconstruct()`
+- X10D: Added `Vector3.Round([float])`
 - X10D: Added `Vector4.Deconstruct()`
+- X10D: Added `Vector4.Round([float])`
 - X10D.Unity: Added `DebugEx`, which mimics `UnityEngine.Debug` while offering more useful primitive drawing methods
 - X10D.Unity: Added `System.Drawing.Color.ToUnityColor()`
 - X10D.Unity: Added `System.Drawing.Color.ToUnityColor32()`
@@ -46,6 +50,7 @@
 - X10D.Unity: Added `Vector2.Deconstruct()`
 - X10D.Unity: Added `Vector2.IsOnLine(LineF)`, `Vector2.IsOnLine(PointF, PointF)`, and `Vector2.IsOnLine(Vector2, Vector2)`
 - X10D.Unity: Added `Vector2Int.IsOnLine(LineF)`, `Vector2Int.IsOnLine(PointF, PointF)`, `Vector2Int.IsOnLine(Vector2, Vector2)`, and `Vector2Int.IsOnLine(Vector2Int, Vector2Int)`
+- X10D.Unity: Added `Vector2.Round([float])`
 - X10D.Unity: Added `Vector2.ToSystemPointF()`
 - X10D.Unity: Added `Vector2.ToSystemSizeF()`
 - X10D.Unity: Added `Vector2Int.Deconstruct()`
@@ -55,12 +60,14 @@
 - X10D.Unity: Added `Vector2Int.WithX()`
 - X10D.Unity: Added `Vector2Int.WithY()`
 - X10D.Unity: Added `Vector3.Deconstruct()`
+- X10D.Unity: Added `Vector3.Round([float])`
 - X10D.Unity: Added `Vector3Int.Deconstruct()`
 - X10D.Unity: Added `Vector3Int.ToSystemVector()`
 - X10D.Unity: Added `Vector3Int.WithX()`
 - X10D.Unity: Added `Vector3Int.WithY()`
 - X10D.Unity: Added `Vector3Int.WithZ()`
 - X10D.Unity: Added `Vector4.Deconstruct()`
+- X10D.Unity: Added `Vector4.Round([float])`
 
 ### Changed
 - X10D.Unity: Obsolesced `Singleton<T>`